feat: migrate command
We added a new CLI command to run database migrations against the
database specified either through the config file or an env variable.

For now there's only one migration to run. It creates a table with a
unique primary key composed of the signing root of a block/attestation
together with the validators public key. The idea is that in the future
this will let us know whether that block has been already been signed.
The migration may still change quite because we're still in a very early
stage regarding slashing protection.

For the migrations we've used the Go tool
github.com/golang-migrate/migrate.

To quickly test the migrations, we added a docker-compose file that
includes the app service and a basic PostgreSQL instance. For now, this
docker-compose is not intended to be used outside of the development
environment, although we'll keep improving it to make installation
easier.