feat: check database connection during server bootstrap
Now that we've started working with the database, we've added a quick
check when starting the HTTP server to verify that we have a connection
to the database. If database is not reachable, we throw an error and
prevent the server from starting, since database access will be
essential for the correct behaviour of the application.