feat: load validators from keystore
We added validator loading from a keystore following the proposal
https://eips.ethereum.org/EIPS/eip-2335 . To do this, we introduced two
new config flags: `keystore_path` and `keystore_password_path`. The
first one is used to specify the path to the validator keys, and the
second one is used to specify the directory with the passphrases if
needed.

Once validator laoding is supported, we also load them before starting
the HTTP server, so in the future we can inject them to the controllers
that will handle the signing.