feat: sign attestation controller
We've added a new endpoint to sign, for now, only attestations. The idea
is keep extending the signer package so that, in addition to the
`SignAttestation` method, we also include `SignBlock` and any other
methods that are needed.

We've also done several refactors to the code's organization. For
example, we moved all the code that builds the HTTP API into its own
function, so we can spin up HTTP servers for automated e2e tests.We've
already written integration tests against the API to validate the
attestation signing logic. Although it currently has some hardcoded
values and connects to the same development database, we'll change that
when we implement slashing protection and stt using the database.

Following a layered architechture, we placed all the logic and
cryptography under `internal` folder, in the `signer` package correctly.
We also gave the `ValidatorKey` struct a method to perform cryptographic
signatures. This keeps all the logic consolidated under the
`internal`folder, and we simply explose the `SignAttestation` method
from our domain to be consumed by the controller, which only needs to
validate the request body.