feat: slashing protection
We add slashing protection using the database. In the first database
migration, we create two tables: `signed_attestations` and
`signed_blocks` to register the signed entities.

In these tables, we'll store a touple with the public BLS address of
signing validator and the signed signing root, so we can check signing
roots already signed and prevent slashing.