feat: support aggregation slots signature
Added support for Aggregation Slot signature type. This type of
signature doesn't require slashing protection, so there is no need to
create any new operation with database.

Simply, as with all signature types, we check whether the received
signingRoot is valid by computing it and then proceed to return the
signature with the specified validator.