feat: support aggregate and proof signatures
Added support for `AGGREGATE_AND_PROOF` signatures. As with all
signature types, we calculate whether we receive the signingRoot and
check it to ensure it is valid.

In this case, slashing protection is not needed for this operation, so
we don't create any table or perform any action to the database.