feat: support sync committee message signature
We add the necessary code to support SYNC_COMMITTEE_MESSAGE signatures,
cryptography, and everything else required. As with all signature types,
we calculate whether we receive the signingRoot and check it to ensure
it is valid. Slashing protection is not necessary for this operation, so
no action will be taken in the database.