feat: network configuration
Instead of having the values needed to compute the domain root hardcoded
for mainnet, we make it configurable, so we can support mainnet, hoodi,
and any other network without problems.

To do this, we've added two json files with the config specification for
mainnet and hoodi, so in these cases the user has not to specify
nothing. For custom networks, we let the user to specify a custom file
with a flag or through config.